+++ package/lib/lo_form.dart
@@ -11,7 +11,11 @@ export 'src/lo_field_state.dart';
 export 'src/lo_form.dart';
 export 'src/lo_form_state.dart';
 export 'src/lo_status.dart';
+export 'src/lo_validator.dart';
 export 'src/types.dart';
 
-// Validation
-export 'validation/lo_validation.dart';
+// Validators
+export 'validators/lo_custom_validator.dart';
+export 'validators/lo_length_validator.dart';
+export 'validators/lo_reg_exp_validator.dart';
+export 'validators/lo_required_validator.dart';